requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bagelML==0.0.21
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
graphlib_backport==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
numpy==2.1.1
openai==1.50.2
overrides==7.7.0
pandas==2.2.3
pydantic==1.10.18
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
style==1.1.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
update==0.0.1
urllib3==2.2.3
Werkzeug==3.0.4
